refactor: 3D 모델 별도 컴포넌트화 #43
