+++ frontend-dashboard/src/components/ThreeDimensionalImage.jsx
@@ -1,32 +1,6 @@
-import { Canvas, useFrame } from "@react-three/fiber";
-import { useGLTF } from "@react-three/drei";
-import { useEffect, useRef } from "react";
+import { Canvas } from "@react-three/fiber";
 import { useThemeStore } from "@/stores/useThemeStore";
-import { useMovementStore } from "@/stores/useMovementStore";
-
-function Model() {
-  const { scene } = useGLTF("/headImage.gltf");
-  const ref = useRef();
-  const isThemeDark = useThemeStore((state) => state.isThemeDark);
-  const { pitch, yaw } = useMovementStore();
-
-  useEffect(() => {
-    scene.traverse((child) => {
-      if (child.isMesh) {
-        child.material.color.set(isThemeDark ? "rgb(140,140,140)" : "white");
-      }
-    });
-  }, [isThemeDark, scene]);
-
-  useFrame(() => {
-    if (ref.current) {
-      ref.current.rotation.x = -(pitch - 0.5) * window.innerWidth * 0.0004;
-      ref.current.rotation.y = -(yaw - 0.5) * window.innerHeight * 0.001;
-    }
-  });
-
-  return <primitive object={scene} ref={ref} position={[0, 0.2, 0]} />;
-}
+import Model3D from "@/components/Model3D";
 
 export default function ThreeDimensionalImage() {
   const isThemeDark = useThemeStore((state) => state.isThemeDark);
@@ -41,7 +15,7 @@ export default function ThreeDimensionalImage() {
       }}
     >
       <hemisphereLight args={["#ffffff", "#444444", 1.5]} position={[0, 20, 0]} />
-      <Model />
+      <Model3D />
     </Canvas>
   );
 }
